WithLogHandler interface, starting integration
